Refactored create-file data in testData file
